requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.1
Django==4.1.2
django-modeladmin-reorder==0.3.1
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
multidict==6.0.2
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
yarl==1.8.1
